changed the way the damage works for quickspells
bc of the colliders on every part of the body I was able to add different amounts of damage to each body part, each spell can now be given a unique amount of damage and its own multiplier to either the head collider or the legs colliders. default is set to 1.5 multiplier for the head and .7 for the legs. whatever tag that body part has will determine how much damage it takes.